refactor(Office365): mutualize metrics
